LinkShortener/requirements.txt
Django==4.0.5
python-dotenv==0.20.0
tzdata==2022.1